relax: move sync backends to backends/ with SyncBackend ABC
- Define SyncBackend(ABC) in core/sync.py with is_configured()/sync()
- Move sync logic from core/sync_github.py → backends/github_sync.GitHubSyncBackend
- Move sync logic from core/sync_gitlab.py → backends/gitlab_sync.GitLabSyncBackend
- All internal helpers are @classmethod; is_configured/sync use @override
- Split GitHubSyncBackend._sync_reviewer_prs into _fetch_reviewer_prs (pure fetch) + _sync_reviewer_prs (cache merge)
- Fix _is_squash_merged glab detection to check for '!' prefix in output
- Update tach.toml to allow backends → core (sync backends need Django models)
- Update tests to import from backends directly, not via core.sync re-exports
- Document SyncBackend ABC contract in BLUEPRINT.md §7 and AGENTS.md
